+++ ICD.Connect.Audio.Misc/BiColorMicButton/BiColorRelayMicButtonDeviceSettings.cs
@@ -1,41 +1,26 @@
-﻿using System;
-using ICD.Common.Utils.Xml;
-using ICD.Connect.Devices;
-using ICD.Connect.Protocol.Ports.DigitalInput;
-using ICD.Connect.Protocol.Ports.IoPort;
+﻿using ICD.Common.Utils.Xml;
+using ICD.Connect.Protocol.Ports.RelayPort;
 using ICD.Connect.Settings.Attributes;
 using ICD.Connect.Settings.Attributes.SettingsProperties;
 
-namespace ICD.Connect.Audio.Misc
+namespace ICD.Connect.Audio.Misc.BiColorMicButton
 {
-	[KrangSettings("BiColorMicButton", typeof(BiColorMicButtonDevice))]
-	public sealed class BiColorMicButtonDeviceSettings : AbstractDeviceSettings
+	[KrangSettings("BiColorRelayMicButton", typeof(BiColorRelayMicButtonDevice))]
+	public sealed class BiColorRelayMicButtonDeviceSettings : AbstractBiColorMicButtonDeviceSettings
 	{
-		private const string BUTTON_INPUT_PORT_ELEMENT = "ButtonInputPort";
 		private const string POWER_OUTPUT_PORT_ELEMENT = "PowerOutputPort";
 		private const string RED_LED_OUTPUT_PORT_ELEMENT = "RedLedOutputPort";
 		private const string GREEN_LED_OUTPUT_PORT_ELEMENT = "GreenLedOutputPort";
-		private const string VOLTAGE_INPUT_PORT_ELEMENT = "VoltageInputPort";
 
-		[OriginatorIdSettingsProperty(typeof(IDigitalInputPort))]
-		public int? ButtonInputPort { get; set; }
-
-		[OriginatorIdSettingsProperty(typeof(IIoPort))]
+		[OriginatorIdSettingsProperty(typeof(IRelayPort))]
 		public int? PowerOutputPort { get; set; }
 
-		[OriginatorIdSettingsProperty(typeof(IIoPort))]
+		[OriginatorIdSettingsProperty(typeof(IRelayPort))]
 		public int? RedLedOutputPort { get; set; }
 
-		[OriginatorIdSettingsProperty(typeof(IIoPort))]
+		[OriginatorIdSettingsProperty(typeof(IRelayPort))]
 		public int? GreenLedOutputPort { get; set; }
 
-		[OriginatorIdSettingsProperty(typeof(IIoPort))]
-		public int? VoltageInputPort { get; set; }
-
-		/// <summary>
-		/// Gets the type of the originator for this settings instance.
-		/// </summary>
-		public override Type OriginatorType { get { return typeof(BiColorMicButtonDevice); } }
 
 		/// <summary>
 		/// Writes property elements to xml.
@@ -45,11 +30,9 @@ namespace ICD.Connect.Audio.Misc
 		{
 			base.WriteElements(writer);
 
-			writer.WriteElementString(BUTTON_INPUT_PORT_ELEMENT, IcdXmlConvert.ToString(ButtonInputPort));
 			writer.WriteElementString(POWER_OUTPUT_PORT_ELEMENT, IcdXmlConvert.ToString(PowerOutputPort));
 			writer.WriteElementString(RED_LED_OUTPUT_PORT_ELEMENT, IcdXmlConvert.ToString(RedLedOutputPort));
 			writer.WriteElementString(GREEN_LED_OUTPUT_PORT_ELEMENT, IcdXmlConvert.ToString(GreenLedOutputPort));
-			writer.WriteElementString(VOLTAGE_INPUT_PORT_ELEMENT, IcdXmlConvert.ToString(VoltageInputPort));
 		}
 
 		/// <summary>
@@ -60,11 +43,9 @@ namespace ICD.Connect.Audio.Misc
 		{
 			base.ParseXml(xml);
 
-			ButtonInputPort = XmlUtils.TryReadChildElementContentAsInt(xml, BUTTON_INPUT_PORT_ELEMENT);
 			PowerOutputPort = XmlUtils.TryReadChildElementContentAsInt(xml, POWER_OUTPUT_PORT_ELEMENT);
 			RedLedOutputPort = XmlUtils.TryReadChildElementContentAsInt(xml, RED_LED_OUTPUT_PORT_ELEMENT);
 			GreenLedOutputPort = XmlUtils.TryReadChildElementContentAsInt(xml, GREEN_LED_OUTPUT_PORT_ELEMENT);
-			VoltageInputPort = XmlUtils.TryReadChildElementContentAsInt(xml, VOLTAGE_INPUT_PORT_ELEMENT);
 		}
 	}
 }